Refactor CursorPosition.col from u16 into usize
Fixes #22

It's *likely* that u16 is big enough, but there's no guarantee that a
text line in a file somewhere or sent from some rogue server won't
exceed that, so downcasting everything to u16 is not very "safe."

Besides, switching to usize simplifies a ton of code since it's the
unit we're typically working with. CursorPosition is not stored *so*
much that a few extra bytes in each will dramatically increase memory
usage.

+++ src/app/prompt.rs
@@ -35,7 +35,7 @@ impl Prompt {
         self.window.inserting = true;
         self.window.cursor = CursorPosition {
             line: 0,
-            col: prompt.width() as u16,
+            col: prompt.width(),
         };
         self.buffer.append(prompt.into());
         self.handle_content_change();
